refactor: rename state files to slice files
