+++ src/components/pages/AddBlogPage/index.tsx
@@ -4,9 +4,9 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Redirect, RouteComponentProps } from 'react-router-dom';
 import styled from 'styled-components';
-import { AddBlogState, addBlogSlice, addBlog } from '../../../redux/slices/add-blog-state';
+import { AddBlogState, addBlogSlice, addBlog } from '../../../redux/slices/add-blog';
 import { AppState } from '../../../redux/app-reducer';
-import { BlogState } from '../../../redux/slices/blog-state';
+import { BlogState } from '../../../redux/slices/blog';
 import Button from '../../atoms/Button/index';
 import Wrapper from '../../atoms/Wrapper/index';
 import AddBlogForm from '../../organisms/AddBlogForm/index';
