+++ src/components/pages/AuthPage/index.tsx
@@ -3,7 +3,7 @@ import 'firebase/auth';
 import React, { Fragment, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Redirect, RouteComponentProps, withRouter } from 'react-router-dom';
-import { UserState, fetchUser } from '../../../redux/slices/user-state';
+import { UserState, fetchUser } from '../../../redux/slices/user';
 import { AppState } from '../../../redux/app-reducer';
 import LoadingView from '../../molecules/LoadingView/index';
 import PageLayout from '../../templates/PageLayout/index';
