+++ src/components/pages/FeedPage/index.tsx
@@ -9,13 +9,13 @@ import { CountType } from '../../../models/consts/count-type';
 import { CountEntity, ItemEntity } from '../../../models/entities';
 import { CountResponse, ItemResponse } from '../../../models/responses';
 import { AppState } from '../../../redux/app-reducer';
-import { FeedState } from '../../../redux/slices/feeds-state';
+import { FeedState } from '../../../redux/slices/feeds';
 import ScrollView from '../../atoms/ScrollView/index';
 import LoadingView from '../../molecules/LoadingView/index';
 import EntryCell, { Count } from '../../organisms/EntryCell/index';
 import { colorsValue } from '../../properties';
 import PageLayout from '../../templates/PageLayout/index';
-import { feedsSlice } from '../../../redux/slices/feeds-state';
+import { feedsSlice } from '../../../redux/slices/feeds';
 
 type CountMap = Map<string, number>;
 type AnimateMap = Map<string, boolean>;
