+++ src/redux/sagas/ga-saga.ts
@@ -1,6 +1,6 @@
 import ReactGA from 'react-ga';
 import { takeLatest } from 'redux-saga/effects';
-import { userSlice } from '../slices/user-state';
+import { userSlice } from '../slices/user';
 
 
 export default function* gaSaga() {
